+++ benchmark/IntMemoryInsert.go
@@ -14,7 +14,7 @@ func InsertInt(param int) (map[int64]Usuario, time.Duration) {
     for i := 0; i < param; i++ {
         str := "name" 
         str2 := "dir" 
-        m[i] = Usuario{ str, str2}
+        m[int64(i)] = Usuario{ str, str2}
     }
     return   m , time.Now().Sub(tiempo1)
 }
